Bug 624842: update build scripts to handle *.abc being located in 'generated' instead of 'core/builtin.abc' and 'shell/shell_toplevel.abc' - additional tweaks to other build files to handle the new location of the generated abc/cpp/h files (r+stejohns)

+++ shell/shell_toplevel.py
@@ -54,7 +54,8 @@ def mv(oldfile, newfile):
     os.remove(oldfile)
 
 def rm(file):
-    os.remove(file)
+    if os.access(file, os.F_OK) == True:
+        os.remove(file)
 
 javacmd = "java -ea -DAS3 -DAVMPLUS -classpath "+classpath
 asc = javacmd+" macromedia.asc.embedding.ScriptCompiler "
